# contracts/openapi/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "SmartQuarter max-gateway"
  "version": "1.1.0"
  "description": "Gateway API with real Identity, Issue and Community services. Session actor and active house are server-owned. House registration requires explicit platform administrator approval. Stable HTTP errors map gRPC status codes."
"servers":
  -
    "url": "http://localhost:18080"
"security":
  -
    "SessionAuth": []
"paths":
  "/api/v1/session/max":
    "post":
      "summary": "Validate MAX and create Redis session"
      "operationId": "post__api_v1_session_max"
      "description": "gRPC: Identity.UpsertMaxUser -> GetUserContext -> GetMembership. Real Identity integration. Users without membership can access house onboarding. Cookie is HttpOnly, Path=/; production Secure + SameSite=None; local Lax."
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Session"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "security": []
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": false
              "properties":
                "init_data":
                  "type": "string"
              "required":
                - "init_data"
            "example":
              "init_data": "<raw MAX initData>"
  "/api/v1/me":
    "get":
      "summary": "Get user context"
      "operationId": "get__api_v1_me"
      "description": "Required role: Any session. gRPC: Identity.GetUserContext. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/UserContext"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
  "/api/v1/session/active-house":
    "post":
      "summary": "Select active membership"
      "operationId": "post__api_v1_session_active-house"
      "description": "Required role: ACTIVE membership of selected house. gRPC: Identity.GetMembership. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "type": "object"
                "additionalProperties": false
                "properties":
                  "active_house_id":
                    "type": "string"
                    "format": "uuid"
                  "role":
                    "type": "string"
                "required":
                  - "active_house_id"
                  - "role"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": false
              "properties":
                "house_id":
                  "type": "string"
                  "format": "uuid"
              "required":
                - "house_id"
  "/api/v1/session/logout":
    "post":
      "summary": "Revoke session"
      "operationId": "post__api_v1_session_logout"
      "description": "Required role: Any session. "
      "responses":
        "204":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/Empty"
  "/api/v1/uploads":
    "post":
      "summary": "Create direct S3 upload"
      "operationId": "post__api_v1_uploads"
      "description": "Required role: ACTIVE member. gRPC: Issue.CreateUpload. "
      "responses":
        "201":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Upload"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/UploadRequest"
            "example":
              "filename": "photo.png"
              "mime_type": "image/png"
              "size_bytes": 1234
  "/api/v1/uploads/{id}/complete":
    "post":
      "summary": "Verify uploaded object"
      "operationId": "post__api_v1_uploads_id_complete"
      "description": "Required role: ACTIVE member. gRPC: Issue.CompleteUpload. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Attachment"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/Empty"
  "/api/v1/attachments/{id}/download-url":
    "get":
      "summary": "Get signed private download URL"
      "operationId": "get__api_v1_attachments_id_download-url"
      "description": "Required role: ACTIVE member. gRPC: Issue.GetAttachmentDownloadURL. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/DownloadURL"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
  "/api/v1/issues":
    "post":
      "summary": "Create issue"
      "operationId": "post__api_v1_issues"
      "description": "Required role: ACTIVE member. gRPC: Issue.CreateIssue. "
      "responses":
        "201":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Issue"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
        -
          "name": "Idempotency-Key"
          "in": "header"
          "schema":
            "type": "string"
            "format": "uuid"
          "description": "Optional. Scope: user + active house. Key binds method, path and compact JSON bytes (key order significant). Same request replays success for 24h; changed payload 409 IDEMPOTENCY_KEY_REUSED; concurrent/ambiguous outcome 409 IDEMPOTENCY_IN_PROGRESS. Reconcile before creating a new key; no exactly-once guarantee."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/CreateIssue"
            "example":
              "category": "INFRASTRUCTURE"
              "description": "Broken lighting"
              "location_text": "Entrance 1"
              "attachment_ids":
                - "55555555-5555-4555-8555-555555555555"
    "get":
      "summary": "List house issues"
      "operationId": "get__api_v1_issues"
      "description": "Required role: ACTIVE member. gRPC: Issue.ListIssues. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/IssueList"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "page_size"
          "in": "query"
          "schema":
            "type": "integer"
            "default": 20
            "minimum": 1
            "maximum": 100
        -
          "name": "page_token"
          "in": "query"
          "schema":
            "type": "string"
          "description": "Opaque token; empty next_page_token means end."
        -
          "name": "status"
          "in": "query"
          "schema":
            "type": "array"
            "items":
              "$ref": "#/components/schemas/IssueStatus"
          "style": "form"
          "explode": true
          "description": "Repeated status parameter; comma-separated values also accepted."
  "/api/v1/issues/{id}":
    "get":
      "summary": "Get issue details"
      "operationId": "get__api_v1_issues_id"
      "description": "Required role: ACTIVE member. gRPC: Issue.GetIssue. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/IssueDetails"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
  "/api/v1/issues/{id}/confirm":
    "post":
      "summary": "Confirm neighbor issue"
      "operationId": "post__api_v1_issues_id_confirm"
      "description": "Required role: ACTIVE member. gRPC: Issue.ConfirmIssue. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Confirmation"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/Empty"
  "/api/v1/issues/{id}/status":
    "patch":
      "summary": "Change issue status"
      "operationId": "patch__api_v1_issues_id_status"
      "description": "Required role: CHAIRMAN or ADMIN. gRPC: Issue.UpdateIssueStatus. "
      "responses":
        "200":
          "description": "Success"
          "headers":
            "X-Request-Id":
              "schema":
                "type": "string"
                "format": "uuid"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Issue"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
        "429":
          "$ref": "#/components/responses/Error429"
        "500":
          "$ref": "#/components/responses/Error500"
        "502":
          "$ref": "#/components/responses/Error502"
        "503":
          "$ref": "#/components/responses/Error503"
        "504":
          "$ref": "#/components/responses/Error504"
        "408":
          "$ref": "#/components/responses/Error408"
      "parameters":
        -
          "name": "X-Request-Id"
          "in": "header"
          "required": false
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "Origin"
          "in": "header"
          "required": true
          "schema":
            "type": "string"
            "format": "uri"
          "description": "Exact origin from TRUSTED_ORIGINS. Missing or foreign Origin is rejected (CSRF protection)."
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "type": "object"
              "additionalProperties": false
              "properties":
                "new_status":
                  "$ref": "#/components/schemas/IssueStatus"
              "required":
                - "new_status"
...
"components":
  "securitySchemes":
    "SessionAuth":
      "type": "apiKey"
      "in": "cookie"
      "name": "sq_session"
    "MaxWebhookSecret":
      "type": "apiKey"
      "in": "header"
      "name": "X-Max-Bot-Api-Secret"
  "schemas":
...
    "Session":
      "type": "object"
      "additionalProperties": false
      "properties":
        "user_context":
          "$ref": "#/components/schemas/UserContext"
        "expires_at":
          "type": "string"
          "format": "date-time"
      "required":
        - "user_context"
        - "expires_at"
...
    "Empty":
      "type": "object"
      "additionalProperties": false
      "properties": {}
      "required": []
...
  "responses":
    "Error400":
      "description": "Invalid JSON, enum, identifier or pagination"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error401":
      "description": "Missing/invalid/expired session or initData"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error403":
      "description": "Forbidden membership, role or Origin"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error404":
      "description": "Resource not found"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error409":
      "description": "Conflict, invalid transition or idempotency reservation"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error413":
      "description": "Body exceeds 256 KiB"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
    "Error415":
      "description": "application/json required"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
...